Bregman kmeans implemented + (minor) ifdef flags name change
The k-means partitioning function is put in a class and can be used
not only for 2-means clustering, but should work for any value of k.
Moreover, this initializes the kmeans with a Gonzalez analogue to
Bregman clustering. So we can even use this to partition data without
doing kmeans -- just using the Gonzalez algorithm to partition the
data (if we want faster tree construction). This behavior can be
achieved by setting the number of kmeans iteration to 0.

Some simple tests are written to test this class even though this
might need a code review + more testing.