Added a main file which reads in options from the command line.
Plus some more changes to the code to comply with the changes
made to the BregmanBall class (keeping around both left and right
centroids and radii).

The given main file also tests the correctness of the left NN
search for a given data set and a test real data set is also
added.